% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commafy.R
\name{commafy}
\alias{commafy}
\title{Function to comma separate a number}
\usage{
commafy(number)
}
\arguments{
\item{number}{a number}
}
\value{
comma separated string
}
\description{
Function to comma separate a number
}
