% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_path.R
\name{data_path}
\alias{data_path}
\title{Standardize file paths}
\usage{
data_path(..., dir = content_dir)
}
\arguments{
\item{...}{objects or strings fed to a paste0 function}

\item{dir}{a root directory}
}
\value{
string
}
\description{
Create path to defined data directory by defining/re-defining content_dir; option to change directory opening for specific purposes.
}
\examples{
content_dir <- "~/r_proj/cats/"
filename <- "dogs.csv"
data_path("data/output/", filename)
}
