% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commafy.R
\name{name_formatter}
\alias{name_formatter}
\title{Name formatter
Function to recode Takoma Park, Montgomery County, PG County, and Maryland}
\usage{
name_formatter(df)
}
\arguments{
\item{df}{ACS dataframe with name column containing Takoma Park, Montgomery County, and/or Maryland or PG County}
}
\value{
Dataframe with location column for Takoma Park, Montgomery County, Prince George's County, or Maryland.
}
\description{
Name formatter
Function to recode Takoma Park, Montgomery County, PG County, and Maryland
}
