% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commafy.R
\name{name_num_formatter}
\alias{name_num_formatter}
\title{Name/number formatter
Applies name and number formatter functions to one dataframe}
\usage{
name_num_formatter(df, round_dig = 2, arrange_col = NULL)
}
\arguments{
\item{df}{Dataframe name/num_formatter functions applied to}

\item{round_dig}{Number of digits to round numeric columns to}

\item{arrange_col}{Column to arrange data by}
}
\value{
Dataframe with rounded numeric columns and location column
}
\description{
Name/number formatter
Applies name and number formatter functions to one dataframe
}
