% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commafy.R
\name{num_formatter}
\alias{num_formatter}
\title{Number formatter
Rounds all numeric columns to 2 decimal places}
\usage{
num_formatter(df, round_dig = 2, arrange_col = NULL)
}
\arguments{
\item{df}{Dataframe to round numeric columns in}

\item{round_dig}{Number of digits to round numeric columns to}

\item{arrange_col}{Column to arrange data by}
}
\value{
Dataframe with numeric columns rounded, arranged by numeric column if identified
}
\description{
Number formatter
Rounds all numeric columns to 2 decimal places
}
