% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_round.R
\name{pct_round}
\alias{pct_round}
\title{Divide numerator by denominator and round to specified digits}
\usage{
pct_round(numerator, denominator, n = 1)
}
\arguments{
\item{numerator}{Numerator in calcualting percentage.}

\item{denominator}{Denominator in calculating percentage.}

\item{n}{the number of digits in the field}
}
\value{
number
}
\description{
Divide numerator by denominator and round to specified digits
}
