% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_acs_tp.R
\name{process_df_tp}
\alias{process_df_tp}
\title{Process ACS data - Takoma Park specific
Groups lightly-processed tidycensus/ACS dataframe, and identifies statistical-significance of values comparing Takoma Park to Montgomery County and Maryland.}
\usage{
process_df_tp(
  df,
  group_cols,
  overall_cols,
  name_col,
  bind_overall = NULL,
  signif_cols = list(Overall = "signif_overall", MC = "signif_mont", MD =
    "signif_maryland"),
  root_df = NULL
)
}
\arguments{
\item{df}{Lightly-processed ACS dataframe with estimate and MOE column.}

\item{group_cols}{Columns to group ACS dataframe by.}

\item{overall_cols}{The set of columns to calculate overall results for. E.g., if you want to see how age groups differ from overall rates of access, you would enter the computer access column as the group col.}

\item{name_col}{Base-name of column to store values in for grouped dataframe.}

\item{bind_overall}{Whether to add overall results as a row to the processed dataframe. Should enter name of column with groups comparing against (e.g., if comparing ages against overall, should enter age).}

\item{signif_cols}{Named list with value representing name of the column with significance values, and names representing how significance should be described in the new column (e.g., "Montgomery County" = "signif_mont" would be for a signif_mont column describing the significance of differences with Montgomery County, and Montgomery County is how it would be represented in the new column).}

\item{root_df}{If the df has already been processed more, the base-dataframe from which the df was processed.}
}
\value{

}
\description{
Process ACS data - Takoma Park specific
Groups lightly-processed tidycensus/ACS dataframe, and identifies statistical-significance of values comparing Takoma Park to Montgomery County and Maryland.
}
