% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_creator.R
\name{pull_creator}
\alias{pull_creator}
\title{function to make pull functions}
\usage{
pull_creator(df_result, filter_col, default_pull = NULL, ...)
}
\arguments{
\item{df_result}{a dataframe you want filtered by one or multiple filter columns and to pull values from}

\item{filter_col}{one or multiple columns to filter the dataframe by}

\item{default_pull}{a default column to pull from the filtered data.}

\item{...}{additional argument options to the returned function}
}
\value{
a function to filter the dataframe by a character vector equal in length to the number of filter columns. returns values based on the filter columns/values provided; typically, will be a vector of length 1
}
\description{
function to make pull functions
}
