% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_format.R
\name{sentence_format}
\alias{sentence_format}
\title{Format character vector as sentence}
\usage{
sentence_format(char_vector, separator = ",")
}
\arguments{
\item{char_vector}{Vector to separate into a sentence format.}

\item{separator}{What to separate items in the character vector by.}
}
\value{

}
\description{
Function to format character vector as list. Can define different separator
}
