% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsfuncts.R
\name{signif_mont_tp}
\alias{signif_mont_tp}
\title{Function to compare statistical significance of results for Takoma Park to Montgomery County and Maryland}
\usage{
signif_mont_tp(df, join_col, name_pct, moe_pct = pct_moe)
}
\arguments{
\item{df}{Processed ACS dataframe with results for Takoma Park, Maryland, and Montgomery County stored in a location column.}

\item{join_col}{Column by which join restuls for Montgomery County/Maryland back to original dataframe.}

\item{name_pct}{Name of percentage column you're comparing values for.}

\item{moe_pct}{Name of moe for percentage column.}
}
\value{
A dataframe with critical values and statistical significance values comparing Takoma Park against Maryland and Montgomery County
}
\description{
Function to compare statistical significance of results for Takoma Park to Montgomery County and Maryland
}
