% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot_title.R
\name{subplot_title}
\alias{subplot_title}
\title{Function to add a subplot title to a plotly subplot}
\usage{
subplot_title(plot, title)
}
\arguments{
\item{plot}{plotly subplot}

\item{title}{title of subplot}
}
\value{
plotly with subtitle
}
\description{
Function to add a subplot title to a plotly subplot
}
